fraud_platform_imen/requirements.txt
easyocr==1.7.0
matplotlib==3.7.2
mysql_connector==2.2.9
nltk==3.8.1
# --upgrade setuptools wheel
numpy==1.24.0
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.76
pandas==1.5.2
Pillow==9.3.0
plotly==5.16.1
bcrypt==4.0.1
scikit_learn==1.2.2
spacy==3.6.1
streamlit==1.16.0
streamlit_camera_input_live==0.2.0
streamlit_card==0.0.61
tqdm==4.65.0
streamlit-extras
